apps/api/requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4.46
alembic>=1.10.0
pydantic>=1.10.7
passlib==1.7.4
bcrypt==3.2.2
python-jose>=3.3.0
python-dotenv>=1.0.0
